Update file structure and template engine
