+++ package.json
@@ -4,17 +4,21 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "start":"node app.js",
+    "start": "node app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Amit",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.20.1",
-    "ejs": "^3.1.8",
-    "dotenv": "^16.0.1",
+    "body-parser": "^1.20.2",
+    "config": "3.3.1",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "hbs": "^4.2.0",
     "mongo": "^0.1.0",
-    "mongoose": "^6.8.0"
+    "mongoose": "^6.10.0",
+    "node": "^19.6.1",
+    "nodeman": "^1.1.2",
+    "validator": "^13.9.0"
   }
 }
