+++ templates/partials/footer.hbs
@@ -1,147 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="description" content="Information" />
-    <meta name="keywords" content="HTML, CSS, JavaScript, VANTA JS, CodeChef, MMMUT, Chapter" />
-    <meta name="author" content="Team Codechef MMMUT" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>CodeChef MMMUT Chapter | Add</title>
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <!--Main CSS-->
-    <link rel="stylesheet" href="../css/style.css">
-    <!-- Team CSS -->
-    <link rel="stylesheet" href="../css/contact.css">
-    <!--fontawesome-->
-    <script src="https://kit.fontawesome.com/332a215f17.js" crossorigin="anonymous"></script>
-    <!--google Fonts-->
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet">
-    <!--Animate.css--->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.0.0/animate.min.css" />
-    <link href="https://unpkg.com/aos@2.3.1/dist/aos.css" rel="stylesheet" />
-    <!-- Favicon -->
-    <link rel="icon" href="/css/img/favicon.ico" type="image/png" />
-</head>
-
-<body onload="myfunction()">
-    <div id="load"></div>
-    <nav class="navbar navbar-expand-lg navbar-dark position-fixed w-100 d-flexbox" style="z-index: 1000">
-        <div class="container-fluid">
-            <div class="navbar-brand">
-                <a href="index.html"><img src="/css/img/logo.png" class="logo" alt="cc-chapter-logo" /></a>
-                <a href="http://www.mmmut.ac.in/" target="_blank"><img src="/css/img/MMMUT LOGO.png" class="logo"
-                        alt="mmmut-logo" /></a>
-            </div>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false"
-                aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                <ul class="navbar-nav ms-auto mb-2 mb-lg-0 fs-4">
-                    <li class="nav-item me-2">
-                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/index.html">Home</a>
-                    </li>
-                    <li class="nav-item me-2">
-                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/src/html/alumni.html">Alumni</a>
-                    </li> 
-                    <li class="nav-item me-2">
-                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/src/html/team.html">Our Team</a>
-                    </li>
-                    <li class="nav-item me-2">
-                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter-verificationhome.onrender.com/">Verification
-                            Page</a>
-                    </li>
-                     
-                    <li class="nav-item me-2">
-                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/src/html/contact.html">Contact Us</a>
-                    </li>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
-    <button class="btn-to-top">
-        <i class="fa fa-arrow-up fa-2x"></i>
-    </button>
-    
-    <section class="p-5 text-center" id="contact" style="background-size: cover; background-position: center center;">
-        <section class="page" id="Key-velidation">
-
-            <!-- Section heading -->
-            <h3 class="text-center font-weight-bold text-white mt-5 mb-3">Welcome To Admin-Area</h3>
-            <!-- Section description -->
-            <h4 class="text-center font-weight-normal text-white mx-auto mb-4 pb-2">Provide us your Information</h4>
-
-            <form class="mb-5 mx-md-5" action="/Add" method="POST">
-
-                <div class="row">
-                    <div class="col-md-12 mb-4">
-
-                        <!-- Name -->
-                        <input type="text" id="name" class="form-control" placeholder="Name" name="name">
-
-                    </div>
-                </div> 
-                <div class="row">   
-                    <div class="col-md-6 mb-4">
-
-                        <!-- Roll No. -->
-                        <input id="rollno" type="number" class="form-control" placeholder="Rollno" name="rollno"
-                            required>
-
-                    </div>
-                
-                    <div class="col-md-6 mb-4">
-
-                        <!-- Year -->
-                        <input id="year" type="text" class="form-control" placeholder="Year" name="year" required>
-
-                    </div>
-                </div>
-
-                <div class="row">
-                    <div class="col-md-12 mb-4">
-
-                        <!-- Role -->
-                        <input type="text" id="Role" class="form-control" placeholder="Role" name="role" required>
-
-                    </div>
-                </div>
-
-                <div class="row">
-                    <div class="col-md-12 mb-4">
-
-                        <!-- Password -->
-                        <input type="password" id="Password" class="form-control" placeholder="Password" name="Password"
-                            required>
-
-                    </div>
-                    <div class="col-md-12">
-
-                        <div class="form-group mb-4">
-                            <textarea class="form-control rounded" id="message" rows="3" placeholder="Message"
-                                name="message"></textarea>
-                        </div>
-
-                        <div class="text-center">
-                            <button type="submit" class="btn bg-primary">Submit</button>
-                        </div>
-
-                    </div>
-                </div>
-
-            </form>
-
-        </section>
-    </section>
-
-
-
 
     <!--Content starts-->
     <div class="content flex">
@@ -266,7 +122,3 @@
 
     <script type="text/javascript" src="../js/particles.js"></script>
     <script type="text/javascript" src="../js/app.js"></script>
-
-</body>
-
-</html>
