+++ templates/partials/headlinks.hbs
@@ -1,0 +1,24 @@
+ <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="description" content="Information" />
+  <meta name="keywords" content="HTML, CSS, JavaScript, VANTA JS, CodeChef, MMMUT, Chapter" />
+  <meta name="author" content="Team Codechef MMMUT" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>CodeChef MMMUT Chapter | Verification Page</title>
+  <!-- Bootstrap CSS -->
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+    integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+  <!--Main CSS-->
+  <link rel="stylesheet" href="../css/style.css">
+  <!-- <link rel="stylesheet" href="waves.css"> -->
+  <link rel="stylesheet" href="../css/contact.css">
+  <!--fontawesome-->
+  <script src="https://kit.fontawesome.com/332a215f17.js" crossorigin="anonymous"></script>
+  <!--google Fonts-->
+  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet">
+  <!--Animate.css--->
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.0.0/animate.min.css" />
+  <link href="https://unpkg.com/aos@2.3.1/dist/aos.css" rel="stylesheet" />
+  <!-- Favicon -->
+  <link rel="icon" href="/css/img/favicon.ico" type="image/png" />
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
