+++ templates/partials/navbar.hbs
@@ -1,0 +1,41 @@
+<div id="load"></div>
+    <nav class="navbar navbar-expand-lg navbar-dark position-fixed w-100 d-flexbox" style="z-index: 1000">
+        <div class="container-fluid">
+            <div class="navbar-brand">
+                <a href="index.html"><img src="/css/img/logo.png" class="logo" alt="cc-chapter-logo" /></a>
+                <a href="http://www.mmmut.ac.in/" target="_blank"><img src="/css/img/MMMUT LOGO.png" class="logo"
+                        alt="mmmut-logo" /></a>
+            </div>
+            <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
+                data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false"
+                aria-label="Toggle navigation">
+                <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                <ul class="navbar-nav ms-auto mb-2 mb-lg-0 fs-4">
+                    <li class="nav-item me-2">
+                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/index.html">Home</a>
+                    </li>
+                    <li class="nav-item me-2">
+                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/src/html/alumni.html">Alumni</a>
+                    </li> 
+                    <li class="nav-item me-2">
+                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/src/html/team.html">Our Team</a>
+                    </li>
+                    <li class="nav-item me-2">
+                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter-verificationhome.onrender.com/">Verification
+                            Page</a>
+                    </li>
+                     
+                    <li class="nav-item me-2">
+                        <a class="nav-link" aria-current="page" href="https://codechef-mmmut-chapter.netlify.app/src/html/contact.html">Contact Us</a>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </nav>
+
+    <button class="btn-to-top">
+        <i class="fa fa-arrow-up fa-2x"></i>
+    </button>
+
