+++ templates/views/Add.hbs
@@ -1,0 +1,91 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    {{>headlinks}}
+</head>
+
+<body onload="myfunction()">
+
+    {{!-- fetching the navbar using hbs --}}
+  {{>navbar}}
+    
+    <section class="p-5 text-center" id="contact" style="background-size: cover; background-position: center center;">
+        <section class="page" id="Key-velidation">
+
+            <!-- Section heading -->
+            <h3 class="text-center font-weight-bold text-white mt-5 mb-3">Welcome To Admin-Area</h3>
+            <!-- Section description -->
+            <h4 class="text-center font-weight-normal text-white mx-auto mb-4 pb-2">Provide us your Information</h4>
+
+            <form class="mb-5 mx-md-5" action="/Add" method="POST">
+
+                <div class="row">
+                    <div class="col-md-12 mb-4">
+
+                        <!-- Name -->
+                        <input type="text" id="name" class="form-control" placeholder="Name" name="name">
+
+                    </div>
+                </div> 
+                <div class="row">   
+                    <div class="col-md-6 mb-4">
+
+                        <!-- Roll No. -->
+                        <input id="rollno" type="number" class="form-control" placeholder="Rollno" name="rollno"
+                            required>
+
+                    </div>
+                
+                    <div class="col-md-6 mb-4">
+
+                        <!-- Year -->
+                        <input id="year" type="text" class="form-control" placeholder="Year" name="year" required>
+
+                    </div>
+                </div>
+
+                <div class="row">
+                    <div class="col-md-12 mb-4">
+
+                        <!-- Role -->
+                        <input type="text" id="Role" class="form-control" placeholder="Role" name="role" required>
+
+                    </div>
+                </div>
+
+                <div class="row">
+                    <div class="col-md-12 mb-4">
+
+                        <!-- Password -->
+                        <input type="password" id="Password" class="form-control" placeholder="Password" name="Password"
+                            required>
+
+                    </div>
+                    <div class="col-md-12">
+
+                        <div class="form-group mb-4">
+                            <textarea class="form-control rounded" id="message" rows="3" placeholder="Message"
+                                name="message"></textarea>
+                        </div>
+
+                        <div class="text-center">
+                            <button type="submit" class="btn bg-primary">Submit</button>
+                        </div>
+
+                    </div>
+                </div>
+
+            </form>
+
+        </section>
+    </section>
+
+
+{{!-- footer using hbs --}}
+  {{>footer}}
+
+
+</body>
+
+</html>
