+++ templates/views/Delete.hbs
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+     {{>headlinks}}
+</head>
+
+<body onload="myfunction()">
+   
+   {{!-- fetching the navbar using hbs --}}
+   {{>navbar}}
+
+    <!-- <div><h2 class="my-3" style="text-align: center;">Welcome To Admin-Area</h2></div> -->
+    <!-- <hr>
+    <br> -->
+    <section class="p-5 text-center" id="contact" style="background-size: cover; background-position: center center;">
+        <section class="page" id="Key-velidation">
+
+            <!-- Section heading -->
+            <h3 class="text-center font-weight-bold text-white mt-5 mb-3">Delete Data</h3>
+            <!-- Section description -->
+            <h4 class="text-center font-weight-normal text-white mx-auto mb-4 pb-2">Provide us your Information</h4>
+
+            <form class="mb-5 mx-md-5" action="/Delete" method="POST">
+              <div class="row">
+                <div class="col-md-12 mb-4">
+
+                  <!-- Role -->
+                  <input type="text" id="key" class="form-control" placeholder="Key" name="key" required>
+
+              </div>
+
+                <div class="col-md-12 mb-4">
+                  <!-- Subject -->
+                  <input type="password" id="Password" class="form-control" placeholder="Admin-Password"
+                    name="Password">
+                </div>
+              </div>
+              <div class="text-center">
+                <button type="submit" class="btn bg-primary">Go</button>
+              </div>
+            </form>
+
+        </section>
+    </section>
+
+{{!-- footer using hbs --}}
+  {{>footer}}
+
+</body>
+
+</html>
