+++ templates/views/Verificationhome.hbs
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  {{>headlinks}}
+</head>
+
+<body onload="myfunction()">
+  
+  {{!-- fetching the navbar using hbs --}}
+  {{>navbar}}
+  
+    <!--Section: Content-->
+    <section class="p-5 text-center" id="contact" style="background-size: cover; background-position: center center;">
+      <section class="page">
+        
+        {{!-- <!-- Section heading --> --}}
+        <h3 class="text-center font-weight-bold text-white mt-5 mb-3">Verify Yourself</h3>
+        
+        <form class="mb-5 mx-md-5" action="/"method="POST">
+          <div class="row">
+            <div class="col-md-12 mb-4">
+              <!-- Subject -->
+              <input type="text" id="subject" class="form-control" placeholder="Verification key"
+                name="key">
+            </div>
+          </div>
+          <div class="text-center">
+            <button type="submit" class="btn bg-primary">Go</button>
+          </div>
+        </form>
+        
+      </section>
+    </section>
+  </div>
+
+
+  {{!-- footer using hbs --}}
+  {{>footer}}
+</body>
+
+</html>
