+++ templates/views/VerificationusingRollno.hbs
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+     {{>headlinks}}
+</head>
+
+<body onload="myfunction()">
+  {{!-- fetching the navbar using hbs --}}
+  {{>navbar}}
+
+    <!-- <div><h2 class="my-3" style="text-align: center;">Welcome To Admin-Area</h2></div> -->
+    <!-- <hr>
+    <br> -->
+    <section class="p-5 text-center" id="contact" style="background-size: cover; background-position: center center;">
+        <section class="page" id="Key-velidation">
+
+            <!-- Section heading -->
+            <h3 class="text-center font-weight-bold text-white mt-5 mb-3">System-Verification Using Roll-no</h3>
+            <!-- Section description -->
+            <h4 class="text-center font-weight-normal text-white mx-auto mb-4 pb-2">Provide us your Information</h4>
+
+            <form class="mb-5 mx-md-5" action="/VerificationusingRollno" method="POST">
+
+                <div class="row">
+                    <div class="col-md-12 mb-4">
+
+                        <!-- Name -->
+                        <input type="text" id="name" class="form-control" placeholder="Name" name="name" required>
+
+                    </div>
+                    <div class="col-md-12 mb-4">
+
+                        <!-- Roll No. -->
+                        <input id="rollno" type="number" class="form-control" placeholder="Rollno" name="rollno"
+                            required>
+
+                    </div>
+                    
+                </div>
+                    <div class="col-md-12">
+                        <div class="text-center">
+                            <button type="submit" class="btn bg-primary">GO</button>
+                        </div>
+
+                    </div>
+                </div>
+
+            </form>
+
+        </section>
+    </section>
+
+  {{!-- footer using hbs --}}
+  {{>footer}}
+
+</body>
+
+</html>
